LIT_article_bot/requirements.txt
python-telegram-bot[job-queue]
feedparser
schedule
requests

python-dotenv
python-dateutil
ollama
goose3

# Scraping
beautifulsoup4
lxml

# RAG
chromadb
sentence-transformers
